% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate.R
\name{conjugate_binomial}
\alias{conjugate_binomial}
\title{Conjugate Bayesian inference for a proportion}
\usage{
conjugate_binomial(y, n, aprior, bprior)
}
\arguments{
\item{y}{Outcome count for binomial data}

\item{n}{Denominator for  binomial data}

\item{aprior}{Beta prior second shape parameter}
}
\value{
List with Beta posterior shape parameters
}
\description{
Conjugate Bayesian inference for a proportion
}
