% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate.R
\name{conjugate_normal}
\alias{conjugate_normal}
\title{Conjugate Bayesian inference for a normal outcome with known variance}
\usage{
conjugate_normal(ybar, se, mprior, sprior)
}
\arguments{
\item{se}{Observed standard error for this empirical mean}

\item{mprior}{Mean of normal prior for mu}

\item{sprior}{SD of normal prior for mu}

\item{y}{Observed empirical mean of a sample generated from a normal
with unknown mean mu and sampling variance of 1}
}
\value{
List with mean and SD of Normal posterior for mu given this prior and data
}
\description{
Conjugate Bayesian inference for a normal outcome with known variance
}
