% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height.R
\name{height_model_fn}
\alias{height_model_fn}
\title{Function that maps prior data, and age that it refers to, to a sample
from the posterior for mean age.  Priors pv from default_priors}
\usage{
height_model_fn(pdata, agei, ...)
}
\description{
Function that maps prior data, and age that it refers to, to a sample
from the posterior for mean age.  Priors pv from default_priors
}
