% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate.R
\name{loss_conj_binomial}
\alias{loss_conj_binomial}
\title{Squared error loss representing disagreement between elicited
information about a proportion and a "prior data" representation.}
\usage{
loss_conj_binomial(pdata, prior, elicited, loss = "mad")
}
\arguments{
\item{elicited}{List of elicited information about a proportion, with components

\code{quantiles} Vector of points on \link{0,1}

\code{probs} Vector of probabilities that the proportion is less than this quantile}
}
\value{
Sum of squared difference between posterior and elicited probabilities
}
\description{
Firstly this function obtains the posterior resulting from
combining a (typically vague) prior Beta distribution with
"prior data" representing an expert's additional knowledge.
}
\details{
The sum of squared differences between the posterior probabilities
and the elicited probabilities, for a set of quantiles, is then
returned.  TODO other losses
}
