% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate.R
\name{loss_conj_normal}
\alias{loss_conj_normal}
\title{Squared error loss representing disagreement between elicited
information about a normal quantity and a "prior data" representation.}
\usage{
loss_conj_normal(pdata, prior, elicited, loss = "mad")
}
\arguments{
\item{elicited}{List of elicited information about a normally distributed quantitity, with components

\code{quantiles} Vector of points on the real line

\code{probs} Vector of probabilities that the quantity is less than this quantile}
}
\value{
Sum of squared difference between posterior and elicited probabilities
}
\description{
Squared error loss representing disagreement between elicited
information about a normal quantity and a "prior data" representation.
}
