% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_mixture.R
\name{loss_mcmc_mixture_binomial}
\alias{loss_mcmc_mixture_binomial}
\title{Squared error loss representing disagreement between elicited
information about a proportion and a "prior data" representation}
\usage{
loss_mcmc_mixture_binomial(pdata, mixprior, elicited)
}
\description{
Firstly, this function obtains the posterior from combining a vague
prior for a logit probability (represented as a mixture of normal
distributions) with a set of binomial "prior data" representing an
expert's additional knowledge.
}
\details{
Secondly, the loss between the posterior and the elicited information
is obtained.
}
