% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_mixture.R
\name{mcmc_mixture_binomial}
\alias{mcmc_mixture_binomial}
\title{Posterior from combining binomial prior data with a vague prior
approximated by a mixture distribution}
\usage{
mcmc_mixture_binomial(pdata, mixprior)
}
\arguments{
\item{pdata}{Binomial prior data: data frame or list with components
\code{y} and \code{n}.}
}
\value{
An \code{rstan} sampling object as returned by \code{sampling()}.
}
\description{
Posterior from combining binomial prior data with a vague prior
approximated by a mixture distribution
}
