% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_mixture.R
\name{mcmc_mixture_normal}
\alias{mcmc_mixture_normal}
\title{Posterior from combining normal prior data with a vague prior
approximated by a mixture distribution}
\usage{
mcmc_mixture_normal(pdata, mixprior, ...)
}
\arguments{
\item{pdata}{Normal prior data: data frame or list with components
\code{ybar} and \code{se}.}
}
\value{
An \code{rstan} sampling object as returned by \code{sampling()}.
}
\description{
Posterior from combining normal prior data with a vague prior
approximated by a mixture distribution
}
