% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_normal.R
\name{normal_objective}
\alias{normal_objective}
\title{Objective function for searching for optimal prior data that
minimises loss between posterior and elicited information}
\usage{
normal_objective(
  par,
  elicited,
  mixprior = NULL,
  model_fn = NULL,
  file = NULL,
  ...
)
}
\arguments{
\item{par}{Prior data as vector transformed to real line:
empirical mean and standard error of implicit normal sample}

\item{mixprior}{Required if model_fn not specified}

\item{model_fn}{Function mapping pdata to posterior sample for estimand}
}
\description{
Parameterised in \code{optim} style
}
