% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_binomial.R
\name{pdata_to_sam_binomial}
\alias{pdata_to_sam_binomial}
\title{Calculate loss for a range of proposed binomial data y, n}
\usage{
pdata_to_sam_binomial(
  pdata,
  model_fn = NULL,
  mixprior = NULL,
  elicited,
  ecdf_x = NULL,
  loss = "mad"
)
}
\arguments{
\item{pdata}{data frame with one row for each proposed prior dataset}

\item{model_fn}{Function with one argument, a 1-row data frame or
list with components y and n, and returning a sample from the
posterior given just this dataset and no other data.  This sample
is to be equated with the elicited information.}

\item{mixprior}{TODO document normal mixture prior specification, to be used
if we don't have access to the model. Either model_fn or mixprior should
be supplied.}

\item{elicited}{TODO}
}
\value{
tidy data frame with the loss for each proposed dataset given
the supplied elicited data
}
\description{
Calculate loss for a range of proposed binomial data y, n
}
