% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priordata-package.R
\docType{package}
\name{priordata-package}
\alias{priordata-package}
\alias{priordata}
\title{The 'priordata' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.7. https://mc-stan.org
}
\author{
\strong{Maintainer}: Christopher Jackson \email{chris.jackson@mrc-bsu.cam.ac.uk}

}
