% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priordata_conjugate.R
\name{priordata_conjugate_binomial}
\alias{priordata_conjugate_binomial}
\title{Obtain binomial prior data by conjugacy given elicited quantiles
and a sample from the vague prior}
\usage{
priordata_conjugate_binomial(sample_prior, elicited)
}
\description{
Obtain binomial prior data by conjugacy given elicited quantiles
and a sample from the vague prior
}
