% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priordata_conjugate.R
\name{priordata_conjugate_normal}
\alias{priordata_conjugate_normal}
\title{Obtain normal prior data by conjugacy given elicited quantiles
and a sample from the vague prior}
\usage{
priordata_conjugate_normal(sample_prior, elicited)
}
\description{
Obtain normal prior data by conjugacy given elicited quantiles
and a sample from the vague prior
}
